Refactor: moved admin related tests
